backend/planner/pyproject.toml
[project]
name = "planner"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "stratos-database",
    "boto3>=1.40.8",
    "langfuse>=3.3.4",
    "openai-agents[litellm]>=0.3.0",
    "massive>=2.0.1",
    "pydantic>=2.11.7",
    "pydantic-ai>=1.0.6",
    "python-dotenv>=1.1.1",
    "tenacity>=9.1.2",
]

[tool.uv.sources]
stratos-database = { path = "../database", editable = true }
